+++ src/Model/PhasorDynamics/Bus/BusInfinite.cpp
@@ -4,7 +4,7 @@
 #include <cmath>
 #include <iostream>
 
-#include <PowerSystemData.hpp>
+#include <Model/PowerFlow/PowerSystemData.hpp>
 
 namespace GridKit
 {
