+++ src/Model/PowerElectronics/DistributedGenerator/DistributedGenerator.hpp
@@ -4,7 +4,6 @@
 #define _CAP_HPP_
 
 #include <Model/PowerElectronics/CircuitComponent.hpp>
-#include <PowerSystemData.hpp>
 
 namespace GridKit
 {
