+++ src/Model/PowerElectronics/InductionMotor/InductionMotor.hpp
@@ -4,7 +4,6 @@
 #define _IMOTOR_HPP_
 
 #include <Model/PowerElectronics/CircuitComponent.hpp>
-#include <PowerSystemData.hpp>
 
 namespace GridKit
 {
