+++ src/Model/PowerElectronics/LinearTransformer/LinearTransformer.hpp
@@ -4,7 +4,6 @@
 #define _LTRANS_HPP_
 
 #include <Model/PowerElectronics/CircuitComponent.hpp>
-#include <PowerSystemData.hpp>
 
 namespace GridKit
 {
