+++ src/Model/PowerElectronics/MicrogridLoad/MicrogridLoad.hpp
@@ -4,7 +4,6 @@
 #define _TRANLOAD_HPP_
 
 #include <Model/PowerElectronics/CircuitComponent.hpp>
-#include <PowerSystemData.hpp>
 
 namespace GridKit
 {
