+++ src/Model/PowerElectronics/Resistor/Resistor.hpp
@@ -4,7 +4,6 @@
 #define _RES_HPP_
 
 #include <Model/PowerElectronics/CircuitComponent.hpp>
-#include <PowerSystemData.hpp>
 
 namespace GridKit
 {
