+++ src/Model/PowerElectronics/SynchronousMachine/SynchronousMachine.hpp
@@ -6,7 +6,6 @@
 #include <tuple>
 
 #include <Model/PowerElectronics/CircuitComponent.hpp>
-#include <PowerSystemData.hpp>
 
 namespace GridKit
 {
