+++ src/Model/PowerElectronics/TransmissionLine/TransmissionLine.hpp
@@ -4,7 +4,6 @@
 #define _TRANLINE_HPP_
 
 #include <Model/PowerElectronics/CircuitComponent.hpp>
-#include <PowerSystemData.hpp>
 
 namespace GridKit
 {
