+++ src/Model/PowerElectronics/VoltageSource/VoltageSource.hpp
@@ -4,7 +4,6 @@
 #define _VOSO_HPP_
 
 #include <Model/PowerElectronics/CircuitComponent.hpp>
-#include <PowerSystemData.hpp>
 
 namespace GridKit
 {
