+++ src/Model/PowerFlow/Generator/GeneratorPV.hpp
@@ -5,7 +5,7 @@
 
 #include "GeneratorBase.hpp"
 #include <Model/PowerFlow/ModelEvaluatorImpl.hpp>
-#include <PowerSystemData.hpp>
+#include <Model/PowerFlow/PowerSystemData.hpp>
 
 namespace GridKit
 {
