+++ src/Model/PowerFlow/Generator/GeneratorSlack.hpp
@@ -1,11 +1,10 @@
 
 #pragma once
 
-// #include <Model/PowerFlow/ModelEvaluatorImpl.hpp>
 #include <vector>
 
 #include "GeneratorBase.hpp"
-#include <PowerSystemData.hpp>
+#include <Model/PowerFlow/PowerSystemData.hpp>
 
 namespace GridKit
 {
