+++ src/Model/PowerFlow/MatpowerParser.hpp
@@ -1,0 +1,299 @@
+
+/**
+ * @file FileIO.hpp
+ * @author Slaven Peles <slaven.peles@pnnl.gov>
+ *
+ * Contains definition of a utility for reading lookup tables.
+ *
+ */
+#pragma once
+
+#include <cstdlib>
+#include <fstream>
+#include <iostream>
+#include <memory>
+#include <ostream>
+#include <regex>
+#include <sstream>
+#include <vector>
+
+#include <Model/PowerFlow/PowerSystemData.hpp>
+
+namespace GridKit
+{
+  using namespace GridKit::PowerSystemData;
+
+  static const std::string matlab_syntax_error{
+      "Only a subset of Matlab syntax is supported."
+      "\n\t'=' for assignment must be on the same line as the field, eg "
+      "`mpc.version = '2'`."
+      "\n\tOpen brace ('[') must be on the same line as the field for matrix "
+      "initialization."
+      "\n\tEach row of a matrix must be terminated by a semicolon."};
+
+  std::ostream& logs()
+  {
+#ifndef NDEBUG
+    std::cerr << "[FileIO.hpp]: ";
+    return std::cerr;
+#else
+    static std::ofstream ofs;
+    ofs.setstate(std::ios_base::badbit);
+    return ofs;
+#endif
+  }
+
+  void ltrim(std::string& s)
+  {
+    const std::string nothing = "";
+    s                         = std::regex_replace(s, std::regex("^\\s+"), nothing);
+  }
+
+  void rtrim(std::string& s)
+  {
+    const std::string nothing = "";
+    s                         = std::regex_replace(s, std::regex("\\s+$"), nothing);
+  }
+
+  void trim_matlab_comments(std::string& s)
+  {
+    const std::string nothing = "";
+    s                         = std::regex_replace(s, std::regex("%.+"), nothing);
+  }
+
+  // Retrive MATPOWER component from assignment line.
+  //
+  // For example, the string "   mpc.bus =  [ ... ] % Some comment" will
+  // return the value "bus".
+  std::string getMatPowerComponent(const std::string& line)
+  {
+    logs() << "Getting matpower component from line\n";
+    std::regex  pat("mpc.([a-zA-Z]+)\\s*=.+");
+    std::smatch matches;
+    std::string component;
+    if (std::regex_match(line, matches, pat))
+    {
+      component = matches[1].str();
+    }
+    else
+    {
+      throw std::runtime_error(matlab_syntax_error + "\nGot line " + line);
+    }
+    ltrim(component);
+    rtrim(component);
+    return component;
+  }
+
+  // Ensure that all of the given line has been consumed, and that the only
+  // remaining non-whitespace character left in the line is a semicolon.
+  void checkEndOfMatrixRow(std::istream& is)
+  {
+    std::string rest;
+    is >> rest;
+    ltrim(rest);
+    rtrim(rest);
+    if (rest != ";")
+      throw std::runtime_error(matlab_syntax_error);
+  }
+
+  template <typename RealT = double, typename IdxT = int>
+  void readMatPowerBusRow(const std::string& row, BusData<RealT, IdxT>& br, LoadData<RealT, IdxT>& lr)
+  {
+    logs() << "Parsing MATPOWER bus row\n";
+    std::stringstream is(row);
+    is >> br.bus_i   // Bus ID
+        >> br.type   // Bus type: 1 = PQ, 2 = PV, 3 = ref, 4 = isolated
+        >> lr.Pd     // Active power demand [MW]
+        >> lr.Qd     // Reactive power demand [MVAr]
+        >> br.Gs     // Shunt conductance (MW demanded at V = 1.0 p.u.)
+        >> br.Bs     // Shunt susceptance (MVAr injected at V = 1.0 p.u.)
+        >> br.area   // Area number (>0)
+        >> br.Vm     // Voltage magnitude (p.u.)
+        >> br.Va     // Voltage phase (deg)
+        >> br.baseKV // Base voltage [kV]
+        >> br.zone   // Loss zone number (>0)
+        >> br.Vmax   // Maximum voltage magnitude (p.u.)
+        >> br.Vmin;  // Minimum voltage magnitude (p.u.)
+
+    lr.bus_i = br.bus_i;
+
+    // std::cout << br.str();
+    // logs() << "Read BusData with the following values:\n" << br.str();
+    // return br;
+  }
+
+  template <typename RealT = double, typename IdxT = int>
+  void readMatPowerGenRow(GenData<RealT, IdxT>& gr, std::string& row)
+  {
+    logs() << "Parsing MATPOWER gen row\n";
+    std::stringstream is(row);
+    is >> gr.bus >> gr.Pg >> gr.Qg >> gr.Qmax >> gr.Qmin >> gr.Vg >> gr.mBase
+        >> gr.status >> gr.Pmax >> gr.Pmin >> gr.Pc1 >> gr.Pc2 >> gr.Qc1min
+        >> gr.Qc1max >> gr.Qc2min >> gr.Qc2max >> gr.ramp_agc >> gr.ramp_10
+        >> gr.ramp_30 >> gr.ramp_q >> gr.apf;
+    checkEndOfMatrixRow(is);
+  }
+
+  template <typename RealT = double, typename IdxT = int>
+  void readMatPowerBranchRow(BranchData<RealT, IdxT>& br, std::string& row)
+  {
+    logs() << "Parsing MATPOWER branch row\n";
+    std::stringstream is(row);
+    is >> br.fbus >> br.tbus >> br.r >> br.x >> br.b >> br.rateA >> br.rateB
+        >> br.rateC >> br.ratio >> br.angle >> br.status >> br.angmin
+        >> br.angmax;
+    checkEndOfMatrixRow(is);
+  }
+
+  template <typename RealT = double, typename IdxT = int>
+  void readMatPowerGenCostRow(GenCostData<RealT, IdxT>& gcr, std::string& row)
+  {
+    logs() << "Parsing MATPOWER gen cost row\n";
+    // Ensure last character is semicolon.
+    rtrim(row);
+    if (row[row.size() - 1] != ';')
+      throw std::runtime_error(matlab_syntax_error + "\nGot line " + row);
+
+    std::stringstream is(row);
+    is >> gcr.kind >> gcr.startup >> gcr.shutdown >> gcr.n;
+
+    for (RealT r; is >> r;)
+    {
+      gcr.rest.push_back(r);
+    }
+  }
+
+  template <typename RealT = double, typename IdxT = int>
+  void readMatPowerVersion(SystemModelData<RealT, IdxT>& mp, std::string& line)
+  {
+    logs() << "Parsing matpower version\n";
+    std::regex  pat("mpc\\.version\\s*=\\s*'([0-9])';");
+    std::smatch matches;
+    if (std::regex_match(line, matches, pat))
+    {
+      mp.version = matches[1].str();
+    }
+    else
+    {
+      throw std::runtime_error(matlab_syntax_error + "\nGot line '" + line + "'");
+    }
+  }
+
+  template <typename RealT = double, typename IdxT = int>
+  void readMatPowerBaseMVA(SystemModelData<RealT, IdxT>& mp, std::string& line)
+  {
+    std::regex  pat("mpc\\.baseMVA\\s*=\\s*([0-9]+);");
+    std::smatch matches;
+    if (std::regex_match(line, matches, pat))
+    {
+      std::string s = matches[1];
+      mp.baseMVA    = std::atof(s.c_str());
+    }
+    else
+    {
+      throw std::runtime_error(matlab_syntax_error + "\nGot line '" + line + "'");
+    }
+  }
+
+  template <typename RealT = double, typename IdxT = int>
+  void readMatPowerFile(SystemModelData<RealT, IdxT>& mp, std::string& filename)
+  {
+    std::ifstream ifs{filename};
+    readMatPower(mp, ifs);
+  }
+
+  template <typename IdxT           = int,
+            typename RealT          = double,
+            std::size_t MaxLineSize = 1028>
+  void readMatPower(SystemModelData<RealT, IdxT>& mp, std::istream& is)
+  {
+    using BusDataT     = BusData<RealT, IdxT>;
+    using GenDataT     = GenData<RealT, IdxT>;
+    using BranchDataT  = BranchData<RealT, IdxT>;
+    using GenCostDataT = GenCostData<RealT, IdxT>;
+    using LoadDataT    = LoadData<RealT, IdxT>;
+
+    for (std::string line; std::getline(is, line);)
+    {
+      // Trim whitespace and remove comments
+      ltrim(line);
+      rtrim(line);
+      logs() << line << "\n";
+      trim_matlab_comments(line);
+
+      // Skip empty lines and comment-only lines
+      if (line.size() == 0)
+        continue;
+
+      // Skip the matlab function declaration
+      if (line.find("function") != std::string::npos)
+        continue;
+
+      // Check for MATPOWER component definitions
+      if (line.find("mpc") != std::string::npos)
+      {
+        const std::string component = getMatPowerComponent(line);
+        logs() << "Got component: '" << component << "'\n";
+        // First, parse matrix components
+        if (component == "bus")
+        {
+          while (std::getline(is, line))
+          {
+            if (line.find("];") != std::string::npos)
+              break;
+            BusDataT  br;
+            LoadDataT lr;
+            readMatPowerBusRow(line, br, lr);
+            mp.bus.push_back(std::move(br));
+            mp.load.push_back(std::move(lr));
+          }
+        }
+        else if (component == "gen")
+        {
+          while (std::getline(is, line))
+          {
+            if (line.find("];") != std::string::npos)
+              break;
+            GenDataT gr;
+            readMatPowerGenRow(gr, line);
+            mp.gen.push_back(gr);
+          }
+        }
+        else if (component == "branch")
+        {
+          while (std::getline(is, line))
+          {
+            if (line.find("];") != std::string::npos)
+              break;
+            BranchDataT br;
+            readMatPowerBranchRow(br, line);
+            mp.branch.push_back(br);
+          }
+        }
+        else if (component == "gencost")
+        {
+          while (std::getline(is, line))
+          {
+            if (line.find("];") != std::string::npos)
+              break;
+            GenCostDataT gcr;
+            readMatPowerGenCostRow(gcr, line);
+            mp.gencost.push_back(gcr);
+          }
+        }
+
+        // Next, parse scalar components
+        else if (component == "version")
+        {
+          readMatPowerVersion(mp, line);
+        }
+        else if (component == "baseMVA")
+        {
+          readMatPowerBaseMVA(mp, line);
+        }
+      }
+    }
+  }
+
+  // } // namespace PowerSystemsData
+} // namespace GridKit
